Added description and licenses, standaridized naming conventions across 3 datasets acccording to the languages for datasets, recipes.
