+++ connectors-endpoints/huggingface-llama2-13b-gptq.json
@@ -2,7 +2,7 @@
     "name": "HuggingFace Llama2 13B GPTQ",
     "connector_type": "huggingface-connector",
     "uri": "",
-    "token": "",
+    "token": "{use your api token}",
     "max_calls_per_second": 1,
     "max_concurrency": 1,
     "model": "",
