+++ connectors-endpoints/openai-gpt4o.json
@@ -2,9 +2,8 @@
   "name": "My OpenAI GPT4o",
   "connector_type": "openai-connector",
   "uri": "",
-  "token": "{your token here}",
+  "token": "{use your token}",
   "max_calls_per_second": 10,
-  "max_concurrency": 10,
   "model": "gpt-4o",
   "params": {
     "timeout": 300,
