+++ connectors-endpoints/together-gemma2-27b.json
@@ -1,11 +1,11 @@
 {
-    "name": "Together Mixtral 8x22b",
+    "name": "Together Gemma 2 - 27B",
     "connector_type": "together-connector",
     "uri": "",
-    "token": "{your token here}",
+    "token": "{use your api token}",
     "max_calls_per_second": 10,
     "max_concurrency": 10,
-    "model": "mistralai/Mixtral-8x22B-Instruct-v0.1",
+    "model": "google/gemma-2-27b-it",
     "params": {
         "allow_retries": true,
         "num_of_retries": 3,
