+++ connectors-endpoints/together-meta-llama-3-1-405b-instruct-connector.json
@@ -1,11 +1,11 @@
 {
-    "name": "Together Gemma 2 - 27B",
+    "name": "Together Llama3.1 405B Instruct",
     "connector_type": "together-connector",
     "uri": "",
-    "token": "your key",
+    "token": "{use your api token}",
     "max_calls_per_second": 10,
     "max_concurrency": 10,
-    "model": "google/gemma-2-27b-it",
+    "model": "meta-llama/Meta-Llama-3.1-405B-Instruct-Turbo",
     "params": {
         "allow_retries": true,
         "num_of_retries": 3,
