+++ connectors-endpoints/together-mixtral-8x22b.json
@@ -1,11 +1,11 @@
 {
-    "name": "Together Llama3.1 405B Instruct",
+    "name": "Together Mixtral 8x22b",
     "connector_type": "together-connector",
     "uri": "",
-    "token": "your token here",
+    "token": "{use your api token}",
     "max_calls_per_second": 10,
     "max_concurrency": 10,
-    "model": "meta-llama/Meta-Llama-3.1-405B-Instruct-Turbo",
+    "model": "mistralai/Mixtral-8x22B-Instruct-v0.1",
     "params": {
         "allow_retries": true,
         "num_of_retries": 3,
