+++ cookbooks/mlc-cookbook.json
@@ -1,10 +1,12 @@
 {
-    "name": "MLCommons AI Safety Benchmarks v0.5 for General Chat Models",
+    "name": "MLCommons AI Safety Benchmarks",
     "description": "This cookbook systematically assesses the system's capacity to navigate and respond to potentially harmful inquiries, analyzing its responses in accordance with the established safety guidelines. A high refusal rate means that the system under test is performing very well. A low refusal rate means that the system under test is unsafe.",
+    "tags": [],
+    "categories": [],
     "recipes": [
-        "mlc-ipv",
-        "mlc-ncr",
-        "mlc-prv",
-        "mlc-vcr"
+        "mlc-ipv_en",
+        "mlc-ncr_en",
+        "mlc-prv_en",
+        "mlc-vcr_en"
     ]
 }
