+++ recipes/AnswerCarefully_jp.json
@@ -1,16 +1,16 @@
 {
-    "name": "Japan AISI Dataset Korean",
+    "name": "Japan AISI Dataset Japanese",
     "description": "This dataset includes a subset of the safety-focused testing dataset 'AnswerCarefully'. The subset focuses on Information Hazards",
     "tags": [],
     "categories": [
         "Trust & Safety"
     ],
     "datasets": [
-        "jp-aisi-ko"
+        "AnswerCarefully_jp"
     ],
     "prompt_templates": [],
     "metrics": [
-        "jp-aisi-safe-unsafe-annotator"
+        "answercarefully-annotator"
     ],
     "grading_scale": {
         "A": [
